Rename Elements of 'comment' Member
Change 'highlight', 'signature', 'affix', 'modified', 'authType', and 'authToken'; to 'what', 'who', 'where', 'when', 'why', and 'how' respectively. Edits were also made to the notation example.

This is to simplify the process of reading, and writing, an entry. 'why' is intended to be interpreted as "why should I believe this is real?', and 'how' is intended to be interpreted as "how can you prove to me it’s real?".