Add Extra Notation-Examples in docs/notes
Create the directory "docs/notes" and place several notation-examples within it. The notation-examples were all generated from the same source-material.

This intends to introduce field-examples on how to use Answer-Notation. More notes will be made and added to this directory over time.

"docs/notes" should be the default directory for answer-notation entries in any repository. I separated my notes by topic and named my files after the questions, but this is not required.