Add 'group' Element to the Notation
Append 'group'-Element to the end of the Notation. Update "Composition"-section and Notation-Example to reflect this change.

'group' is intended to allow gathering of semantically-relevant notes together, from both a human and machine perspective.

Before this inclusion, there was no way to symbolize in-data that different notes were taken together. This is useful if trying to group notes from a particular lecture, session, or workload.