Add 'subgroup' Element to Notation
Append 'subgroup'-array Element, along with member-elements 'tag' and 'hint', to the notation. Update the "Composition"-section and Notation-Example to reflect these changes.

'subgroup' is intended to allow for distributed flagging of notes. This is useful for when categorizing notes using concepts, ideas, and theory.

'subgroup' can crawl ALL notes making it an effective tool for targeted research. Compared to 'group' which, can only crawl a subset of notes and, is an effective tool for structural preservation.